Remove interpreter folder
