--- pennylane/compiler/python_compiler/transforms/api/interpreter/__init__.py
@@ -1,22 +1,0 @@
-# Copyright 2025 Xanadu Quantum Technologies Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""xDSL transforms interpreter API."""
-
-from .transform_functions import TransformFunctionsExt
-from .transform_interpreter_pass import TransformInterpreterPass
-
-__all__ = [
-    "TransformFunctionsExt",
-    "TransformInterpreterPass",
-]
